+++ src/status.rs
@@ -22,7 +22,20 @@ pub struct Location {
 pub struct Spacefed {
     pub spacenet: bool,
     pub spacesaml: bool,
-    pub spacephone: bool,
+    pub spacephone: Option<bool>,
+}
+
+impl Spacefed {
+    fn verify(&self, version: StatusBuilderVersion) -> Result<(), String> {
+        if version == StatusBuilderVersion::V14 {
+            if self.spacephone.is_some() {
+                return Err("spacefed.spacephone key was removed".into());
+            }
+        } else if self.spacephone.is_none() {
+            return Err("spacefed.spacephone must be present".into());
+        }
+        Ok(())
+    }
 }
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
@@ -202,6 +215,7 @@ pub struct Status {
     // SpaceAPI internal usage
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cache: Option<Cache>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
     pub issue_report_channels: Vec<IssueReportChannel>,
 
     // Mutable data
@@ -241,7 +255,7 @@ impl Status {
     }
 }
 
-#[derive(Debug, Clone, Eq, PartialEq)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
 enum StatusBuilderVersion {
     V0_13,
     V14,
@@ -383,9 +397,26 @@ impl StatusBuilder {
         };
 
         let contact = self.contact.ok_or("contact missing")?;
+        if let Some(spacefed) = &self.spacefed {
+            spacefed.verify(self.version)?;
+        }
 
-        if self.version == StatusBuilderVersion::V14 && contact.jabber.is_some() {
-            return Err("Jabber key under contact was renamed to xmpp".into());
+        if self.version == StatusBuilderVersion::V14 {
+            if contact.jabber.is_some() {
+                return Err("jabber key under contact was renamed to xmpp".into());
+            }
+            if contact.google.is_some() {
+                return Err("google key under contact was removed".into());
+            }
+            if self.radio_show.is_some() {
+                return Err("radio_show key was removed".into());
+            }
+
+            if !self.issue_report_channels.is_empty() {
+                return Err("issue_report_channels key was removed".into());
+            }
+        } else if self.issue_report_channels.is_empty() {
+            return Err("issue_report_channels must not be empty".into());
         }
 
         Ok(Status {
@@ -453,7 +484,6 @@ mod test {
             .url("foobar")
             .location(Location::default())
             .contact(Contact::default())
-            .add_issue_report_channel(IssueReportChannel::Email)
             .build()
             .unwrap();
         assert_eq!(
@@ -464,7 +494,7 @@ mod test {
                 space: "foo".into(),
                 logo: "bar".into(),
                 url: "foobar".into(),
-                issue_report_channels: vec![IssueReportChannel::Email],
+                issue_report_channels: vec![],
                 ..Status::default()
             }
         );
@@ -480,7 +510,47 @@ mod test {
                 jabber: Some("jabber".into()),
                 ..Contact::default()
             })
-            .add_issue_report_channel(IssueReportChannel::Email)
+            .build();
+        assert!(status.is_err());
+    }
+
+    #[test]
+    fn test_builder_v14_fail_on_google() {
+        let status = StatusBuilder::v14("foo")
+            .logo("bar")
+            .url("foobar")
+            .location(Location::default())
+            .contact(Contact {
+                google: Some(GoogleContact::default()),
+                ..Contact::default()
+            })
+            .build();
+        assert!(status.is_err());
+    }
+
+    #[test]
+    fn test_builder_v14_fail_on_radio_show() {
+        let status = StatusBuilder::v14("foo")
+            .logo("bar")
+            .url("foobar")
+            .location(Location::default())
+            .contact(Contact::default())
+            .add_radio_show(RadioShow::default())
+            .build();
+        assert!(status.is_err());
+    }
+
+    #[test]
+    fn test_builder_v14_fail_on_spacephone() {
+        let status = StatusBuilder::v14("foo")
+            .logo("bar")
+            .url("foobar")
+            .location(Location::default())
+            .contact(Contact::default())
+            .spacefed(Spacefed {
+                spacephone: Some(true),
+                ..Spacefed::default()
+            })
             .build();
         assert!(status.is_err());
     }
@@ -516,7 +586,10 @@ mod test {
             .url("foobar")
             .location(Location::default())
             .contact(Contact::default())
-            .spacefed(Spacefed::default())
+            .spacefed(Spacefed {
+                spacephone: Some(false),
+                ..Spacefed::default()
+            })
             .feeds(Feeds::default())
             .add_project("spaceapi-rs")
             .add_cam("cam1")
@@ -531,7 +604,13 @@ mod test {
         assert_eq!(status.url, "foobar");
         assert_eq!(status.location, Location::default());
         assert_eq!(status.contact, Contact::default());
-        assert_eq!(status.spacefed, Some(Spacefed::default()));
+        assert_eq!(
+            status.spacefed,
+            Some(Spacefed {
+                spacephone: Some(false),
+                ..Spacefed::default()
+            })
+        );
         assert_eq!(status.feeds, Some(Feeds::default()));
         assert_eq!(status.projects, Some(vec!["spaceapi-rs".to_string()]));
         assert_eq!(status.cam, Some(vec!["cam1".to_string(), "cam2".to_string()]));
@@ -567,6 +646,7 @@ mod test {
             .location(Location::default())
             .contact(Contact::default())
             .add_extension("aaa", Value::Array(vec![Value::Null, Value::from(42)]))
+            .add_issue_report_channel(IssueReportChannel::Email)
             .build()
             .unwrap();
         let serialized = to_string(&status).unwrap();
@@ -581,12 +661,13 @@ mod test {
             .url("c")
             .location(Location::default())
             .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
             .build();
         assert!(status.is_ok());
         assert_eq!(
             &to_string(&status.unwrap()).unwrap(),
             "{\"api\":\"0.13\",\"space\":\"a\",\"logo\":\"b\",\"url\":\"c\",\
-             \"location\":{\"lat\":0.0,\"lon\":0.0},\"contact\":{},\"issue_report_channels\":[],\
+             \"location\":{\"lat\":0.0,\"lon\":0.0},\"contact\":{},\"issue_report_channels\":[\"email\"],\
              \"state\":{\"open\":null}}"
         );
     }
@@ -598,6 +679,7 @@ mod test {
             .url("c")
             .location(Location::default())
             .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
             .add_extension("aaa", Value::String("xxx".into()))
             .add_extension("bbb", Value::Array(vec![Value::Null, Value::from(42)]))
             .build();
@@ -605,7 +687,7 @@ mod test {
         assert_eq!(
             &to_string(&status.unwrap()).unwrap(),
             "{\"api\":\"0.13\",\"space\":\"a\",\"logo\":\"b\",\"url\":\"c\",\
-             \"location\":{\"lat\":0.0,\"lon\":0.0},\"contact\":{},\"issue_report_channels\":[],\
+             \"location\":{\"lat\":0.0,\"lon\":0.0},\"contact\":{},\"issue_report_channels\":[\"email\"],\
              \"state\":{\"open\":null},\"ext_aaa\":\"xxx\",\"ext_bbb\":[null,42]}"
         );
     }
@@ -617,6 +699,7 @@ mod test {
             .url("foobar")
             .location(Location::default())
             .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
             .build()
             .unwrap();
         status.spacefed = Some(Spacefed::default());
@@ -640,6 +723,7 @@ mod test {
             .url("c")
             .location(Location::default())
             .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
             .add_extension("aaa", Value::String("xxx".into()))
             .add_extension("ext_aaa", Value::String("yyy".into()))
             .add_extension("bbb", Value::Null)